ai-runtime/requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
python-multipart
langchain
langchain-community
psutil
stable-diffusion-cpp-python
requests
